ports: Move `LockedFreePort` from postgresql-testing for reuse.
