+++ libs-scala/ports/src/main/scala/com/digitalasset/ports/PortLock.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.testing.postgresql
+package com.daml.ports
 
 import java.io.RandomAccessFile
 import java.nio.channels.{
@@ -12,9 +12,7 @@ import java.nio.channels.{
 }
 import java.nio.file.{Files, Path, Paths}
 
-import com.daml.ports.Port
-
-private[postgresql] object PortLock {
+object PortLock {
 
   // We can't use `sys.props("java.io.tmpdir")` because Bazel changes this for each test run.
   // For this to be useful, it needs to be shared across concurrent runs.
