+++ libs-scala/ports/src/test/suite/scala/com/digitalasset/ports/LockedFreePortSpec.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.testing.postgresql
+package com.daml.ports
 
 import org.scalatest.{Matchers, WordSpec}
